+++ src/test/java/ActionsTests.java
@@ -1,3 +1,4 @@
+import PageModel.ActionsPage;
 import PageModel.LoginPage;
 import org.openqa.selenium.By;
 import org.openqa.selenium.Keys;
@@ -13,22 +14,17 @@ import java.util.List;
 public class ActionsTests extends BaseTest {
     @Test
     public void playSongTest() {
-        // hover over in clickPlayBtn
+        ActionsPage actionsPage = new ActionsPage(driver);
         LoginPage loginPage = new LoginPage(driver);
         loginPage.login("demo@class.com", "te$t$tudent");
-        clickFooterPlayBtn();
-        Assert.assertTrue(pauseBtnExists());
-
-        // Comparing numbers of elements example
+        actionsPage.clickFooterPlayBtn();
+        Assert.assertTrue(actionsPage.pauseBtnExists());
         List<WebElement> songs = driver.findElements(By.cssSelector("[data-test='song-card']"));
-
         int songsNumberBefore = songs.size();
         System.out.println(songsNumberBefore);
         // Just an example: here we would add or delete an element but we didn't
         int songsNumberAfter = songs.size();
         System.out.println(songsNumberAfter);
-
-        // Soft assert example
         SoftAssert softAssert = new SoftAssert();
         softAssert.assertTrue(songsNumberBefore == songsNumberAfter,
                 "=== Songs number before should be equal songs number after ===");
@@ -39,93 +35,39 @@ public class ActionsTests extends BaseTest {
 
     @Test
     public void renamePlaylist() {
-        // double click
+    ActionsPage actionsPage =new ActionsPage(driver);
         String playlistName = "Summer songs";
         LoginPage loginPage = new LoginPage(driver);
         loginPage.login("demo@class.com", "te$t$tudent");
-        doubleClickChoosePlaylist();
-        enterPlaylistName(playlistName);
-        String newName = getPlaylistName();
+        actionsPage.doubleClickChoosePlaylist();
+        actionsPage.enterPlaylistName(playlistName);
+        String newName = actionsPage.getPlaylistName();
         Assert.assertEquals(playlistName, newName, "=== PlaylistNames expected to be equals ===");
     }
 
     @Test
     public void playSongFromListTest() {
+        ActionsPage actionsPage =new ActionsPage(driver);
         LoginPage loginPage = new LoginPage(driver);
         loginPage.login("demo@class.com", "te$t$tudent");
-        goToAllSongs();
-        // right click on first song
-        rightClickOnFirstSong();
-        // click play button
-        clickPlayBtn();
-        // verify
-        Assert.assertTrue(isEqualizerDisplayed());
+        actionsPage.goToAllSongs();
+        actionsPage.rightClickOnFirstSong();
+       actionsPage.clickPlayBtn();
+        Assert.assertTrue(actionsPage.isEqualizerDisplayed());
     }
 
-    private boolean isEqualizerDisplayed() {
-        return driver.findElement(By.cssSelector("[data-testid='sound-bar-play']")).isDisplayed();
-    }
-
-
-    private void clickPlayBtn() {
-        WebElement playBtn = driver.findElement(By.cssSelector(".playback"));
-        playBtn.click();
-    }
-
-    private void rightClickOnFirstSong() {
-        WebElement firstSong = driver.findElement(By.cssSelector(".song-item"));
-        Actions actions = new Actions(driver);
-        actions.contextClick(firstSong).perform();
-    }
-
-    public void clickFooterPlayBtn() {
-        Actions action = new Actions(driver);
-        WebElement playBtn = driver.findElement(By.cssSelector("[data-testid='play-btn']"));
-        action
-                .moveToElement(playBtn)
-                .perform();
-        playBtn.click();
-    }
-
-    public boolean pauseBtnExists() {
-        return driver.findElement(By.cssSelector("[data-testid='pause-btn']")).isDisplayed();
-    }
-
-
-    private void goToAllSongs() {
-        wait.until(ExpectedConditions.elementToBeClickable(By.cssSelector(".songs"))).click();
-    }
-
-
-    public void doubleClickChoosePlaylist() {
-        WebElement playlistElement = wait.until(ExpectedConditions.
-                elementToBeClickable(By.cssSelector(".playlist:nth-child(3)")));
-        Actions actions = new Actions(driver);
-        actions.doubleClick(playlistElement).perform();
-    }
-
-    public void enterPlaylistName(String name) {
-        WebElement playlistInputField = driver.findElement(By.cssSelector("input[name='name']"));
-        playlistInputField.sendKeys(Keys.HOME, Keys.chord(Keys.SHIFT, Keys.END), name);
-        playlistInputField.sendKeys(Keys.ENTER);
-    }
-
-    private String getPlaylistName() {
-        WebElement playlistElement = wait.until(ExpectedConditions.
-                visibilityOfElementLocated(By.cssSelector(".playlist:nth-child(3)>a")));
-        String name = playlistElement.getText();
-        return name;
-    }
 
     @Test
     public void countSongsInPlaylist() {
+        ActionsPage actionsPage =new ActionsPage(driver);
         LoginPage loginPage = new LoginPage(driver);
         loginPage.login("demo@class.com", "te$t$tudent");
-        WebElement playlist = wait.until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector(".playlist:nth-child(4)")));
-        playlist.click();
-        List<WebElement> songs = driver.findElements(By.cssSelector("#playlistWrapper .song-item"));
+        actionsPage.playListClick();
+        List<WebElement> songs = actionsPage.getWebElements();
         int number = songs.size();
         Assert.assertEquals(number, 1); // can fail, depends on current number. This is just an example
 
     }
+
+
 }
