+++ src/test/java/PageModel/ActionsPage.java
@@ -1,0 +1,79 @@
+package PageModel;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.Keys;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.interactions.Actions;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+
+import java.util.List;
+
+public class ActionsPage extends Basepage{
+    public ActionsPage(WebDriver givenDriver) {
+        super(givenDriver);
+    }
+    public boolean isEqualizerDisplayed() {
+        return driver.findElement(By.cssSelector("[data-testid='sound-bar-play']")).isDisplayed();
+    }
+
+
+    public void clickPlayBtn() {
+        WebElement playBtn = driver.findElement(By.cssSelector(".playback"));
+        playBtn.click();
+    }
+
+    public void rightClickOnFirstSong() {
+        WebElement firstSong = driver.findElement(By.cssSelector(".song-item"));
+        Actions actions = new Actions(driver);
+        actions.contextClick(firstSong).perform();
+    }
+
+    public void clickFooterPlayBtn() {
+        Actions action = new Actions(driver);
+        WebElement playBtn = driver.findElement(By.cssSelector("[data-testid='play-btn']"));
+        action
+                .moveToElement(playBtn)
+                .perform();
+        playBtn.click();
+    }
+
+    public boolean pauseBtnExists() {
+        return driver.findElement(By.cssSelector("[data-testid='pause-btn']")).isDisplayed();
+    }
+
+
+    public void goToAllSongs() {
+        wait.until(ExpectedConditions.elementToBeClickable(By.cssSelector(".songs"))).click();
+    }
+
+
+    public void doubleClickChoosePlaylist() {
+        WebElement playlistElement = wait.until(ExpectedConditions.
+                elementToBeClickable(By.cssSelector(".playlist:nth-child(3)")));
+        Actions actions = new Actions(driver);
+        actions.doubleClick(playlistElement).perform();
+    }
+
+    public void enterPlaylistName(String name) {
+        WebElement playlistInputField = driver.findElement(By.cssSelector("input[name='name']"));
+        playlistInputField.sendKeys(Keys.HOME, Keys.chord(Keys.SHIFT, Keys.END), name);
+        playlistInputField.sendKeys(Keys.ENTER);
+    }
+
+    public String getPlaylistName() {
+        WebElement playlistElement = wait.until(ExpectedConditions.
+                visibilityOfElementLocated(By.cssSelector(".playlist:nth-child(3)>a")));
+        String name = playlistElement.getText();
+        return name;
+    }
+    public List<WebElement> getWebElements() {
+        List<WebElement> songs = driver.findElements(By.cssSelector("#playlistWrapper .song-item"));
+        return songs;
+    }
+
+    public void playListClick() {
+        WebElement playlist = wait.until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector(".playlist:nth-child(4)")));
+        playlist.click();
+    }
+}
