+++ src/test/java/PageModel/SongsPage.java
@@ -13,8 +13,7 @@ public class SongsPage extends Basepage {
         super(givenDriver);
     }
 
-//    @FindBy(css ="[title='Play or resume']" )
-  //     WebElement playPauseBtn;
+
    By playPauseBtn =By.cssSelector("[title='Play or resume']");
 
     By playBtnelement = By.cssSelector("[data-testid='pause-btn']");
